requirements.txt
mysql-connector-python==9.2.0
tqdm==4.67.1
